[devmgr] unify memfs minfs vfs.[ch] helpers in devmgr
next patch will add the same vfs.[ch] to minfs, then
subsequent patch(es) will move them into a common library

made common trace functions for debug in trace.h
created vfs-devmgr.c to hold devmgr-specific vfs code
created memfs-private.h to declaere devmgr-specific vnode & ifcs
reordered include files as necessary to get local-> system-> std
