+++ system/core/devmgr/vfs-boot.c
@@ -2,8 +2,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "vfs.h"
 #include "dnode.h"
+#include "memfs-private.h"
+#include "vfs.h"
 
 #include <magenta/listnode.h>
