+++ system/core/devmgr/vfs-device.c
@@ -2,24 +2,24 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "vfs.h"
+
 #include "dnode.h"
 #include "devmgr.h"
-
-#include <magenta/listnode.h>
+#include "device-internal.h"
+#include "memfs-private.h"
+#include "vfs.h"
 
 #include <ddk/device.h>
+
+#include <magenta/listnode.h>
+#include <magenta/processargs.h>
+#include <magenta/syscalls.h>
 
 #include <mxio/debug.h>
 #include <mxio/vfs.h>
 
-#include <magenta/processargs.h>
-#include <magenta/syscalls.h>
-
 #include <stdlib.h>
 #include <string.h>
-
-#include "device-internal.h"
 
 
 #define MXDEBUG 0
