+++ system/core/devmgr/vfs.c
@@ -2,59 +2,21 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+#include "trace.h"
 #include "vfs.h"
-#include "devmgr.h"
-#include "dnode.h"
-
-#include <mxio/debug.h>
-#include <mxio/dispatcher.h>
-#include <mxio/io.h>
-#include <mxio/remoteio.h>
-
-#include <ddk/device.h>
-
-#include <magenta/listnode.h>
-
-#include <magenta/device/device.h>
-#include <magenta/device/devmgr.h>
-#include <magenta/processargs.h>
-#include <magenta/syscalls.h>
 
 #include <fcntl.h>
-#include <stdio.h>
-#include <stdlib.h>
 #include <string.h>
-#include <threads.h>
+#include <unistd.h>
 
-#define MXDEBUG 0
+#include <mxio/dispatcher.h>
+#include <mxio/remoteio.h>
 
-#define DEBUG_TRACK_NAMES 1
+struct vnode {
+    VNODE_BASE_FIELDS
+};
 
-mtx_t vfs_lock = MTX_INIT;
-
-static list_node_t vfs_iostate_list = LIST_INITIAL_VALUE(vfs_iostate_list);
-static mtx_t vfs_iostate_lock = MTX_INIT;
-
-void track_vfs_iostate(vfs_iostate_t* ios, const char* fn) {
-#if DEBUG_TRACK_NAMES
-    if (fn) {
-        ios->fn = strdup(fn);
-    }
-#endif
-    mtx_lock(&vfs_iostate_lock);
-    list_add_tail(&vfs_iostate_list, &ios->node);
-    mtx_unlock(&vfs_iostate_lock);
-}
-
-void untrack_vfs_iostate(vfs_iostate_t* ios) {
-    mtx_lock(&vfs_iostate_lock);
-    list_delete(&ios->node);
-    mtx_unlock(&vfs_iostate_lock);
-#if DEBUG_TRACK_NAMES
-    free((void*)ios->fn);
-    ios->fn = NULL;
-#endif
-}
+uint32_t __trace_bits;
 
 // Starting at vnode vn, walk the tree described by the path string,
 // until either there is only one path segment remaining in the string
@@ -80,7 +42,7 @@ mx_status_t vfs_walk(vnode_t* vn, vnode_t** out,
         }
         if (vn->flags & V_FLAG_REMOTE) {
             // remote filesystem mount, caller must resolve
-            xprintf("vfs_walk: vn=%p name='%s' (remote)\n", vn, path);
+            trace(WALK, "vfs_walk: vn=%p name='%s' (remote)\n", vn, path);
             *out = vn;
             *pathout = path;
             if (oldvn == NULL) {
@@ -96,7 +58,7 @@ mx_status_t vfs_walk(vnode_t* vn, vnode_t** out,
             // traverse to the next segment
             len = nextpath - path;
             nextpath++;
-            xprintf("vfs_walk: vn=%p name='%.*s' nextpath='%s'\n", vn, (int)len, path, nextpath);
+            trace(WALK, "vfs_walk: vn=%p name='%.*s' nextpath='%s'\n", vn, (int)len, path, nextpath);
             r = vn->ops->lookup(vn, &vn, path, len);
             if (oldvn) {
                 // release the old vnode, even if there was an error
@@ -109,7 +71,7 @@ mx_status_t vfs_walk(vnode_t* vn, vnode_t** out,
             path = nextpath;
         } else {
             // final path segment, we're done here
-            xprintf("vfs_walk: vn=%p name='%s' (local)\n", vn, path);
+            trace(WALK, "vfs_walk: vn=%p name='%s' (local)\n", vn, path);
             if (oldvn == NULL) {
                 // returning our original vnode, need to upref it
                 vn_acquire(vn);
@@ -123,7 +85,7 @@ mx_status_t vfs_walk(vnode_t* vn, vnode_t** out,
 
 mx_status_t vfs_open(vnode_t* vndir, vnode_t** out, const char* path,
                      const char** pathout, uint32_t flags, uint32_t mode) {
-    xprintf("vfs_open: path='%s' flags=%d mode=%x\n", path, flags, mode);
+    trace(VFS, "vfs_open: path='%s' flags=%d\n", path, flags);
     mx_status_t r;
     if ((r = vfs_walk(vndir, &vndir, path, &path)) < 0) {
         return r;
@@ -141,7 +103,6 @@ mx_status_t vfs_open(vnode_t* vndir, vnode_t** out, const char* path,
 
     size_t len = strlen(path);
     vnode_t* vn;
-
     if (flags & O_CREAT) {
         if ((r = vndir->ops->create(vndir, &vn, path, len, mode)) < 0) {
             if ((r == ERR_ALREADY_EXISTS) && (!(flags & O_EXCL))) {
@@ -178,7 +139,6 @@ mx_status_t vfs_open(vnode_t* vndir, vnode_t** out, const char* path,
         // opening a vnode.
         vn_release(vn);
         if (r < 0) {
-            xprintf("vn open r = %d", r);
             return r;
         }
         if (flags & O_TRUNC) {
@@ -188,6 +148,7 @@ mx_status_t vfs_open(vnode_t* vndir, vnode_t** out, const char* path,
             }
         }
     }
+    trace(VFS, "vfs_open: vn=%p\n", vn);
     *pathout = "";
     *out = vn;
     return NO_ERROR;
@@ -208,13 +169,13 @@ static mx_status_t txn_handoff_rename(mx_handle_t srv, mx_handle_t rh,
     return mxrio_txn_handoff(srv, rh, &msg);
 }
 
-mx_status_t vfs_rename(vnode_t* vn, const char* oldpath, const char* newpath,
+mx_status_t vfs_rename(vnode_t* vndir, const char* oldpath, const char* newpath,
                        mx_handle_t rh) {
     vnode_t* oldparent, *newparent;
     mx_status_t r, r_old, r_new;
-    if ((r_old = vfs_walk(vn, &oldparent, oldpath, &oldpath)) < 0) {
+    if ((r_old = vfs_walk(vndir, &oldparent, oldpath, &oldpath)) < 0) {
         return r_old;
-    } else if ((r_new = vfs_walk(vn, &newparent, newpath, &newpath)) < 0) {
+    } else if ((r_new = vfs_walk(vndir, &newparent, newpath, &newpath)) < 0) {
         vn_release(oldparent);
         return r_new;
     } else if (r_old != r_new) {
@@ -226,7 +187,7 @@ mx_status_t vfs_rename(vnode_t* vn, const char* oldpath, const char* newpath,
 
     if (r_old == 0) {
         // Local filesystem
-        r = vn->ops->rename(oldparent, newparent, oldpath, strlen(oldpath),
+        r = vndir->ops->rename(oldparent, newparent, oldpath, strlen(oldpath),
                             newpath, strlen(newpath));
     } else {
         // Remote filesystem.
@@ -257,95 +218,28 @@ mx_status_t vfs_fill_dirent(vdirent_t* de, size_t delen,
     return sz;
 }
 
-ssize_t vfs_do_ioctl(vnode_t* vn, uint32_t op, const void* in_buf,
-                     size_t in_len, void* out_buf, size_t out_len) {
-    switch (op) {
-    case IOCTL_DEVICE_WATCH_DIR: {
-        if ((out_len != sizeof(mx_handle_t)) || (in_len != 0)) {
-            return ERR_INVALID_ARGS;
-        }
-        if (vn->dnode == NULL) {
-            // not a directory
-            return ERR_WRONG_TYPE;
-        }
-        vnode_watcher_t* watcher;
-        if ((watcher = calloc(1, sizeof(vnode_watcher_t))) == NULL) {
-            return ERR_NO_MEMORY;
-        }
-        mx_handle_t h;
-        if (mx_channel_create(0, &h, &watcher->h) < 0) {
-            free(watcher);
-            return ERR_NO_RESOURCES;
-        }
-        memcpy(out_buf, &h, sizeof(mx_handle_t));
-        mtx_lock(&vfs_lock);
-        list_add_tail(&vn->watch_list, &watcher->node);
-        mtx_unlock(&vfs_lock);
-        xprintf("new watcher vn=%p w=%p\n", vn, watcher);
-        return sizeof(mx_handle_t);
-    }
-    case IOCTL_DEVMGR_MOUNT_FS: {
-        if ((in_len != 0) || (out_len != sizeof(mx_handle_t))) {
-            return ERR_INVALID_ARGS;
-        }
-        mx_handle_t h0, h1;
-        mx_status_t status;
-        if ((status = mx_channel_create(0, &h0, &h1)) < 0) {
-            return status;
-        }
-        if ((status = vfs_install_remote(vn, h1)) < 0) {
-            mx_handle_close(h0);
-            mx_handle_close(h1);
-            return status;
-        }
-        memcpy(out_buf, &h0, sizeof(mx_handle_t));
-        return sizeof(mx_handle_t);
-    }
-    case IOCTL_DEVMGR_UNMOUNT_NODE: {
-        return vfs_uninstall_remote(vn);
-    }
-    default:
-        return vn->ops->ioctl(vn, op, in_buf, in_len, out_buf, out_len);
-    }
-}
-
 void vn_acquire(vnode_t* vn) {
+    trace(REFS, "acquire vn=%p ref=%u\n", vn, vn->refcount);
     vn->refcount++;
 }
 
+// TODO(orr): figure out x-system panic
+#define panic(fmt...) do { fprintf(stderr, fmt); __builtin_trap(); } while (0)
+
 void vn_release(vnode_t* vn) {
+    trace(REFS, "release vn=%p ref=%u\n", vn, vn->refcount);
     if (vn->refcount == 0) {
-        printf("vn %p: ref underflow\n", vn);
-        panic();
+        panic("vn %p: ref underflow\n", vn);
     }
     vn->refcount--;
     if (vn->refcount == 0) {
+        trace(VFS, "vfs_release: vn=%p\n", vn);
         vn->ops->release(vn);
     }
 }
 
-void vfs_dump_handles(void) {
-    vfs_iostate_t* ios;
-    mtx_lock(&vfs_iostate_lock);
-    list_for_every_entry (&vfs_iostate_list, ios, vfs_iostate_t, node) {
-        printf("obj %p '%s'\n", ios->vn, ios->fn ? ios->fn : "???");
-    }
-    mtx_unlock(&vfs_iostate_lock);
-}
-
-void vfs_notify_add(vnode_t* vn, const char* name, size_t len) {
-    xprintf("devfs: notify vn=%p name='%.*s'\n", vn, (int)len, name);
-    vnode_watcher_t* watcher;
-    vnode_watcher_t* tmp;
-    list_for_every_entry_safe (&vn->watch_list, watcher, tmp, vnode_watcher_t, node) {
-        mx_status_t status;
-        if ((status = mx_channel_write(watcher->h, 0, name, len, NULL, 0)) < 0) {
-            xprintf("devfs: watcher %p write failed %d\n", watcher, status);
-            list_delete(&watcher->node);
-            mx_handle_close(watcher->h);
-            free(watcher);
-        } else {
-            xprintf("devfs: watcher %p notified\n", watcher);
-        }
-    }
+mx_status_t vfs_close(vnode_t* vn) {
+    trace(VFS, "vfs_close: vn=%p\n", vn);
+    mx_status_t r = vn->ops->close(vn);
+    return r;
 }
